+++ src/main/java/woowacourse/shoppingcart/config/LoginInterceptor.java
@@ -4,7 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.servlet.HandlerInterceptor;
-import woowacourse.shoppingcart.exception.unauthorized.UnauthorizedTokenException;
+import woowacourse.shoppingcart.exception.UnauthorizedTokenException;
 import woowacourse.shoppingcart.support.AuthorizationExtractor;
 import woowacourse.shoppingcart.support.JwtTokenProvider;
